==> 2_MLD_Analysis.ipynb <==
import numpy as np
import xarray as xr
import pandas as pd

CMIP5_MLD = xr.open_dataset('Data/CMIP5_spatial_MLD.nc').sortby('model')
CMIP6_MLD = xr.open_dataset('Data/CMIP6_spatial_MLD.nc').sortby('model')

CMIP5ds = xr.open_dataset('Data/CMIP5_time_series_smoothed.nc')
CMIP6ds = xr.open_dataset('Data/CMIP6_time_series_smoothed.nc')

zos_mods5 = CMIP5ds.model.values
zos_mods6 = CMIP6ds.model.values

both_5 = list(set(CMIP5_MLD.model.values)&set(zos_mods5))
both_6 = list(set(CMIP6_MLD.model.values)&set(zos_mods6))

CMIP5_MLD = CMIP5_MLD.sel(model=both_5)
CMIP6_MLD = CMIP6_MLD.sel(model=both_6)

# Labrador Sea
lats_LS = np.arange( 50.5, 64.5)
lons_LS = np.arange(-61.5,-34.5)

lats_GS = np.arange( 65.5, 80.5)
lons_GS = np.arange(-10.5, 15.5)

lats_IS = np.arange( 50.5, 67.5)
lons_IS = np.arange(-34.5,  0.5)

CMIP5_MLD_LS = CMIP5_MLD.sel(lat=lats_LS, lon=lons_LS)
CMIP5_MLD_GS = CMIP5_MLD.sel(lat=lats_GS, lon=lons_GS)
CMIP5_MLD_IS = CMIP5_MLD.sel(lat=lats_IS, lon=lons_IS)

data_CMIP5 = {'Model': CMIP5_MLD.model.values, 'LS_mean':CMIP5_MLD_LS.MLD.mean(dim=['lat','lon']).values, 
              'GS_mean':CMIP5_MLD_GS.MLD.mean(dim=['lat','lon']).values, 'IS_mean': CMIP5_MLD_IS.MLD.mean(dim=['lat','lon']).values, 
              'LS_max': CMIP5_MLD_LS.MLD.max(dim=['lat','lon']).values,  'GS_max':CMIP5_MLD_GS.MLD.max(dim=['lat','lon']).values,
              'IS_max':CMIP5_MLD_IS.MLD.max(dim=['lat','lon']).values }
df5 = pd.DataFrame(data = data_CMIP5)
df5.round(2)
df5.set_index('Model')

df5['region_mean'] = df5[['LS_mean','GS_mean','IS_mean']].idxmax(axis=1)
df5['region_max']  = df5[['LS_max','GS_max','IS_max']].idxmax(axis=1)


df5['GS_mean>170'] = df5['GS_mean']>170
df5['LS_mean>170'] = df5['LS_mean']>170
df5['IS_mean>170'] = df5['IS_mean']>170
df5['GS_mean>160'] = df5['GS_mean']>160
df5['GS_mean>180'] = df5['GS_mean']>180
df5['GS_mean>190'] = df5['GS_mean']>190

df5

df5.to_csv('Results_MLD/Regional_MLD_CMIP5.csv')

CMIP6_MLD_LS = CMIP6_MLD.sel(lat=lats_LS, lon=lons_LS)
CMIP6_MLD_GS = CMIP6_MLD.sel(lat=lats_GS, lon=lons_GS)
CMIP6_MLD_IS = CMIP6_MLD.sel(lat=lats_IS, lon=lons_IS)

data_CMIP6 = {'Model': CMIP6_MLD.model.values, 'LS_mean':CMIP6_MLD_LS.MLD.mean(dim=['lat','lon']).values, 
              'GS_mean':CMIP6_MLD_GS.MLD.mean(dim=['lat','lon']).values, 'IS_mean': CMIP6_MLD_IS.MLD.mean(dim=['lat','lon']).values, 
              'LS_max': CMIP6_MLD_LS.MLD.max(dim=['lat','lon']).values,  'GS_max':CMIP6_MLD_GS.MLD.max(dim=['lat','lon']).values,
              'IS_max':CMIP6_MLD_IS.MLD.max(dim=['lat','lon']).values }
df6 = pd.DataFrame(data = data_CMIP6)
df6.round(2)
df6.set_index('Model')

df6['region_mean'] = df6[['LS_mean','GS_mean','IS_mean']].idxmax(axis=1)
df6['region_max']  = df6[['LS_max','GS_max','IS_max']].idxmax(axis=1)

df6['GS_mean>170'] = df6['GS_mean']>170
df6['LS_mean>170'] = df6['LS_mean']>170
df6['IS_mean>170'] = df6['IS_mean']>170
df6['GS_mean>160'] = df6['GS_mean']>160
df6['GS_mean>180'] = df6['GS_mean']>180
df6['GS_mean>190'] = df6['GS_mean']>190

df6

df6.to_csv('Results_MLD/Regional_MLD_CMIP6.csv')
