==> 4_T-test.ipynb <==
import numpy as np
import xarray as xr
import pandas as pd
import matplotlib.pyplot as plt
from scipy import stats

CMIP5ds = xr.open_dataset('Data/CMIP5_time_series_smoothed.nc')
CMIP6ds = xr.open_dataset('Data/CMIP6_time_series_smoothed.nc')

# Loop over reference periods and scenarios:
ttest_results_total = []
for reference_period in [[1900, 1949], [1986, 2005], [1995, 2014]]:
    
    ref_st, ref_end = reference_period[0], reference_period[1]
    CMIP5ds = CMIP5ds - CMIP5ds.sel(time=slice(ref_st,ref_end+1)).mean(dim='time')
    CMIP6ds = CMIP6ds - CMIP6ds.sel(time=slice(ref_st,ref_end+1)).mean(dim='time')
    
    ref_results = []
    print(f'REFERENCE PERIOD: {ref_st}-{ref_end}')
    for sces in range(3):
        CMIP5 = CMIP5ds.isel(scenario=sces)
        CMIP6 = CMIP6ds.isel(scenario=sces)

        CMIP5 = CMIP5['zos'].dropna('model','any')
        CMIP6 = CMIP6['zos'].dropna('model','any')

        print('Result', CMIP5.scenario.values, CMIP6.scenario.values)
        diff_CMIP5 = CMIP5.sel(time=slice(2081, 2101)).mean(dim='time') - CMIP5.sel(time=slice(ref_st, ref_end)).mean(dim='time')
        print(f'Mean difference CMIP5: {np.mean(diff_CMIP5.values):.2f} cm')
        print('Number of models:', len(diff_CMIP5.values))

        diff_CMIP6 = CMIP6.sel(time=slice(2081, 2101)).mean(dim='time') - CMIP6.sel(time=slice(ref_st, ref_end)).mean(dim='time')
        print(f'Mean difference CMIP6: {np.mean(diff_CMIP6.values):.2f} cm')
        print('Number of models:', len(diff_CMIP6.values),'\n')

        print(f'Difference between the two means: {(np.mean(diff_CMIP5.values)-np.mean(diff_CMIP6.values)):.2f} cm' )
        print(f'p-value: {stats.ttest_ind(diff_CMIP5, diff_CMIP6)[1]:.4f}')
        print(stats.ttest_ind(diff_CMIP5, diff_CMIP6),'\n')

        results = [np.mean(diff_CMIP5.values),np.mean(diff_CMIP6.values),np.mean(diff_CMIP6.values)-np.mean(diff_CMIP5.values),stats.ttest_ind(diff_CMIP5, diff_CMIP6)[1]]

        ref_results.append(results)
    
    ttest_results_total.append(ref_results)

d = {'index': ['mean CMIP5', 'mean CMIP6', 'difference CMIP6 - CMIP5', 'p-value'], '1900-1950': ttest_results_total[0][0], '1986-2005':ttest_results_total[1][0], '1995-2014':ttest_results_total[2][0]}
df_26 =  pd.DataFrame(data=d).set_index('index')  
print('Results for RCP2.6/SSP1-2.6:')
df_26.round(3)

d = {'index': ['mean CMIP5', 'mean CMIP6', 'difference CMIP6 - CMIP5', 'p-value'], '1900-1950': ttest_results_total[0][1], '1986-2005':ttest_results_total[1][1], '1995-2014':ttest_results_total[2][1]}
df_45 =  pd.DataFrame(data=d).set_index('index') 
print('Results for RCP4.5/SSP2-4.5:')
df_45.round(3)

d = {'index': ['mean CMIP5', 'mean CMIP6', 'difference CMIP6 - CMIP5', 'p-value'], '1900-1950': ttest_results_total[0][2], '1986-2005':ttest_results_total[1][2], '1995-2014':ttest_results_total[2][2]}
df_85 =  pd.DataFrame(data=d).set_index('index') 
print('Results for RCP8.5/SSP5-8.5:')
df_85.round(3)
